+++ src/components/Forms/_HelpText/HelpText.js
@@ -1,6 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
+import Paragraph from '../../Typography/Paragraph';
 import { Color } from '../../../shared';
 
 import './help-text.scss';
@@ -8,8 +9,8 @@ import './help-text.scss';
 /**
  * The HelpText component is used within this very same library to render input related help texts
  */
-const HelpText = ({ text, color, className, ...rest }) => (
-  <p className={classNames('bi bi-helptext', `helptext-${color}`, className)} {...rest}>{text}</p>
+const HelpText = ({ text, className, ...rest }) => (
+  <Paragraph className={classNames('bi-helptext', className)} {...rest}>{text}</Paragraph>
 );
 
 HelpText.propTypes = {