+++ styleguidist.config.js
@@ -64,6 +64,7 @@ module.exports = {
           name: 'Forms',
           pagePerSection: true,
           components: () => [
+            'src/components/Forms/Checkbox/Checkbox.js',
             'src/components/Forms/Input/Input.js',
             'src/components/Forms/Select/Select.js',
           ],